feat!: fwhm is a property of the shape
Up to now a peak could contain the 2 properties: 'fwhm' and 'shape'. However fwhm is a property of the shape and therefore fwhm could be at the level of the peak or at the level of the shape. Very confusing.

In this breaking change if you want to specify fwhm you need to specify it at the level of the 'shape'.

An old code that was (gaussian is the default shape kind):
`addPeak({ x: 0, y: 1, fwhm: 0.2 })`

Should therefore be rewritten as
`addPeak({ x: 0, y: 1, shape: { kind: 'gaussian', fwhm: 0.2 } })`

peakWdithFct is now optional (no default value) and by default a shape {kind: 'gaussian', fwhm: 5} will be used.

If `peakWidthFct` is defined it will be the priority to defined `fwhm`.